+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -932,6 +932,18 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       public void testRequiredAnnotationResultsInMissingBindings() {
         run("RequiredAnnotationResultsInMissingBindings.kt");
       }
+
+      @Test
+      @TestMetadata("RequiredConstructorDependencyWinsOptionalDependency.kt")
+      public void testRequiredConstructorDependencyWinsOptionalDependency() {
+        run("RequiredConstructorDependencyWinsOptionalDependency.kt");
+      }
+
+      @Test
+      @TestMetadata("RequiredProviderDependencyWinsOptionalDependency.kt")
+      public void testRequiredProviderDependencyWinsOptionalDependency() {
+        run("RequiredProviderDependencyWinsOptionalDependency.kt");
+      }
     }
 
     @Nested